refactor: move executePgmCallBackTest and executeSourceLineTest to AbstractTest
