+++ rpgJavaInterpreter-core/src/test/kotlin/com/smeup/rpgparser/AbstractTest.kt
@@ -23,14 +23,18 @@ import com.smeup.rpgparser.execution.*
 import com.smeup.rpgparser.interpreter.*
 import com.smeup.rpgparser.jvminterop.JavaSystemInterface
 import com.smeup.rpgparser.parsing.ast.CompilationUnit
+import com.smeup.rpgparser.parsing.facade.SourceReferenceType
+import com.smeup.rpgparser.parsing.parsetreetoast.ParseTreeToAstError
 import com.smeup.rpgparser.rpginterop.DirRpgProgramFinder
 import com.smeup.rpgparser.rpginterop.RpgProgramFinder
 import com.smeup.rpgparser.rpginterop.SourceProgramFinder
 import org.apache.logging.log4j.LogManager
+import org.junit.Assert
 import java.io.File
 import java.io.PrintStream
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
+import kotlin.test.fail
 
 /**
  * This class must be extended from all test classes in order to automatically manage tests using both version
@@ -359,6 +363,158 @@ abstract class AbstractTest {
     }
 
     open fun useCompiledVersion() = false
+
+    /**
+     * This function is used to test the execution of a program and validate the error handling mechanism.
+     * It expects the program to fail and checks if the error events are correctly captured.
+     *
+     * @param pgm The name of the program to be executed.
+     * @param sourceReferenceType The expected type of the source reference (Program or Copy) where the error is expected to occur.
+     * @param sourceId The expected identifier of the source where the error is expected to occur.
+     * @param lines The list of line numbers where the errors are expected.
+     * @param reloadConfig The reload configuration to be used for the execution of the program. Default is null.
+     */
+    protected fun executePgmCallBackTest(
+        pgm: String,
+        sourceReferenceType: SourceReferenceType,
+        sourceId: String,
+        lines: List<Int>,
+        reloadConfig: ReloadConfig? = null
+    ) {
+        val errorEvents = mutableListOf<ErrorEvent>()
+        runCatching {
+            val configuration = Configuration().apply {
+                jarikoCallback.onError = { errorEvent ->
+                    println(errorEvent)
+                    errorEvents.add(errorEvent)
+                }
+                options = Options(debuggingInformation = true)
+                this.reloadConfig = reloadConfig
+            }
+            executePgm(pgm, configuration = configuration)
+        }.onSuccess {
+            Assert.fail("Program must exit with error")
+        }.onFailure {
+            println(it.stackTraceToString())
+            Assert.assertEquals(sourceReferenceType, errorEvents[0].sourceReference!!.sourceReferenceType)
+            Assert.assertEquals(sourceId, errorEvents[0].sourceReference!!.sourceId)
+            Assert.assertEquals(lines.sorted(), errorEvents.map { errorEvent -> errorEvent.sourceReference!!.relativeLine }.sorted())
+        }
+    }
+
+    /**
+     * This function is used to test the execution of a program and validate the error handling mechanism.
+     * It expects the program to fail and checks if the error events are correctly captured.
+     *
+     * @param pgm The name of the program to be executed.
+     * @param sourceReferenceType The expected type of the source reference (Program or Copy) where the error is expected to occur.
+     * @param sourceId The expected identifier of the source where the error is expected to occur.
+     * @param lines The map of lines, number and message, expected.
+     * @param reloadConfig The reload configuration to be used for the execution of the program. Default is null.
+     */
+    protected fun executePgmCallBackTest(
+        pgm: String,
+        sourceReferenceType: SourceReferenceType,
+        sourceId: String,
+        lines: Map<Int, String>,
+        reloadConfig: ReloadConfig? = null
+    ) {
+        val errorEvents = mutableListOf<ErrorEvent>()
+        runCatching {
+            val configuration = Configuration().apply {
+                jarikoCallback.onError = { errorEvent ->
+                    println(errorEvent)
+                    errorEvents.add(errorEvent)
+                }
+                options = Options(debuggingInformation = true)
+                this.reloadConfig = reloadConfig
+            }
+            executePgm(pgm, configuration = configuration)
+        }.onSuccess {
+            Assert.fail("Program must exit with error")
+        }.onFailure {
+            println(it.stackTraceToString())
+            Assert.assertEquals(sourceReferenceType, errorEvents[0].sourceReference!!.sourceReferenceType)
+            Assert.assertEquals(sourceId, errorEvents[0].sourceReference!!.sourceId)
+            val found = errorEvents
+                .associate { errorEvent ->
+                    errorEvent.sourceReference!!.relativeLine to (errorEvent.error as ParseTreeToAstError).message!!
+                }
+                .map {
+                    Pair(it.value, it.contains(lines))
+                }
+            Assert.assertTrue(
+                "Errors doesn't correspond:\n" + found.joinToString(separator = "\n") { it.first },
+                found.size == found.filter { it.second }.size && found.size == lines.size
+            )
+        }
+    }
+
+    private fun Map.Entry<Int, String>.contains(list: Map<Int, String>): Boolean {
+        list.forEach {
+            if (this.value.contains(it.value) && this.key == it.key) {
+                return true
+            }
+        }
+        return false
+    }
+
+    /**
+     * Verify that the sourceLine is properly set in case of error.
+     * ErrorEvent must contain a reference to an absolute line of the source code
+     * @param pgm The name of the program to be executed.
+     * @param throwableConsumer A consumer to handle the throwable, default is empty.
+     * @param additionalProgramFinders A list of additional program finders to be used during the execution, default is empty.
+     * @param reloadConfig The reload configuration to be used during the execution, default is null.
+     * */
+    protected fun executeSourceLineTest(
+        pgm: String,
+        throwableConsumer: (Throwable) -> Unit = {},
+        additionalProgramFinders: List<RpgProgramFinder> = emptyList(),
+        reloadConfig: ReloadConfig? = null
+    ) {
+        val sourceIdToLines = mutableMapOf<String, List<String>>()
+        val errorEvents = mutableListOf<ErrorEvent>()
+        val configuration = Configuration().apply {
+            jarikoCallback.beforeParsing = { it ->
+                sourceIdToLines[MainExecutionContext.getParsingProgramStack().peek().name] = it.lines()
+                it
+            }
+            jarikoCallback.onError = {
+                errorEvents.add(it)
+            }
+            // I set dumpSourceOnExecutionError because I want test also the sourceLine presence in case
+            // of runtime error
+            options = Options(debuggingInformation = true, dumpSourceOnExecutionError = true)
+            this.reloadConfig = reloadConfig
+        }
+        kotlin.runCatching {
+            executePgm(pgm, configuration = configuration, additionalProgramFinders = additionalProgramFinders)
+        }.onSuccess {
+            Assert.fail("$pgm must exit with error")
+        }.onFailure {
+            throwableConsumer(it)
+            if (errorEvents.any { errorEvent -> errorEvent.sourceReference!!.sourceId == "UNKNOWN" }) {
+                val errorEvent = errorEvents.first { errorEvent -> errorEvent.sourceReference!!.sourceId == "UNKNOWN" }
+                error("errorEvent: $errorEvent\nwith sourceId: UNKNOWN is not allowed")
+            }
+            errorEvents.forEach { errorEvent ->
+                // copy is not parsed so, if sourceReference is a copy, I will use the program name
+                val sourceId = if (errorEvent.sourceReference!!.sourceReferenceType == SourceReferenceType.Copy) {
+                    pgm
+                } else {
+                    errorEvent.sourceReference!!.sourceId
+                }
+                val lines = sourceIdToLines[sourceId]!!
+                if (lines[errorEvent.absoluteLine!! - 1] != errorEvent.fragment) {
+                    System.err.println("ErrorEvent: $errorEvent")
+                    System.err.println("Jariko arose an error at this line: ${errorEvent.absoluteLine!!}, fragment: ${errorEvent.fragment}")
+                    System.err.println("But the source code at line: ${errorEvent.absoluteLine!!} is: ${lines[errorEvent.absoluteLine!! - 1]}")
+                    fail()
+                }
+            }
+        }
+    }
 }
 
 fun Configuration.adaptForTestCase(testCase: AbstractTest): Configuration {
